Reduced memory usage of MultiPermutation object.
Since the last entry of each permutation is redundant, we can skip that part for saving date in the MultiDimArrayLong object. The only downside is that at the moment, MultiPermutation cannot hold data when the init vector is such that there is effectively only a single permutation (resulting from a vector of one or more 1s).