Added a simple profiler.
If the flag `ENABLE_PROFILING` is enabled in `te-symbolic` or `te-extended`, the individual tasks are registered and the total CPU time for each task is listed at the end of the program. Should help to identify which are the points in the evaluation which are really taking the most time to compute.